fix 0-init of local ( stack ) buffer.
`buffer_limit` might be bigger than the `MAX_PACKET_SIZE` of the local buffer

usually the `buffer_limit` is ensured during receiving data and it should
not be possible to trigger this stack overwrite with an malicious packet.
the buffer_limit in the default config i at least twice the MAX_PACKET_SIZE
and other not connected in any way to the size of the local buffer.

for correctnes, use the minimum of the packet buffer and the local buffer size.