Fix memory leak with respawning dnstap collector
Once a dnstap collector has crashed or gotten out of sync at some point, it
exits and is respawn, but has the zones content in memory then.  This commit
makes sure that after that happened, a new dnstap collector is forked with each
reload so that the memory will not diverge from the serve childs (which will
cause a gradual memory leak).

TODO: Investigate whether it would be worthwhile to reuse the dt_env and/or the
fd_send and fd_recv when respawning a dnstap collector on reload.