Extra defense against double dnstap logging from Wouter
The service remaining tcp routine is already protected by the set of no more tcp queries on the event handler functions so it should never call dnstap log routines, but this is even more explicit to stop double writes to the logs. If we still want to log from the service remaining tcp, need to set up some pipe for that.