Keep a list of changes from the catalog zone, apply them afterwards
Current behaviour is that all changes are applied whilst the catalog zone is processed. This means that for a zone with a group property, the following tasks are created in order: 1) remove zone, 2) add zone with default pattern, 3) remove zone, 4) add zone with pattern from group property.

This changes that to one operation