Many improvements to the nsd.conf man page
All references to the "database" option and database file name have
been removed, save for one mention of the option, to say that it is
ignored in NSD >= 4.8.0.

All references to "master" zones have been changed to "primary", and all
references to "slave" zones have been changed to "secondary", in line
with RFC 8499.

End-of-line spaces have been deleted from all lines. In some places, the
text has been adjusted a little to read better.