Test package for bad xfr with addzone in one tasklist
This test is to make sure that the effects of other tasks are effectuated when
they are processed together with other tasks in a single tasklist.  The test
does a reload for a single bad update (from xfr) together with an addzone in a
single tasklist. The addzone should be effectuated.