contrib/nsd.openrc.in: new OpenRC service file
This OpenRC service script is modified from the one currently used in
Gentoo Linux to start NSD. The main benefit of having it upstream
(i.e. here) is that several paths like @sbindir@ can be passed in
from the build system. Having it here should also avoid some
duplication of effort.

Currently the script is unusable because the build system is unaware
of it (and therefore doesn't make the necessary substitutions), but
that will change in a subsequent commit.