contrib/nsd-tmpfiles.conf.in: new tmpfiles.d entry for local sockets
Both systemd and OpenRC can make use of tmpfiles.d entries as
described in

  https://www.freedesktop.org/software/systemd/man/latest/tmpfiles.d.html

If the user wants to use a local control socket, then a writable
temporary directory is needed to hold the socket. We cannot know
beforehand what the user's configuration will be, so this commit adds
a tmpfiles.d entry that unconditionally creates @runstatedir@/nsd and
makes it writable by the nsd user.

If the temporary directory turns out not to be needed, a superfluous
one is no big deal. But if the user does want to use a local socket,
then this saves him the trouble of securely creating a directory to
hold it.