Change XDP program load/unload to not require CAP_SYS_ADMIN
This depends on PR xdp-project/xdp-tools 432 to get merged. If the PR
doesn't get merged, we need to resort to setting the environment
variable LIBXDP_SKIP_DISPATCHER=1, because unloading an XDP program
using the dispatcher requires CAP_SYS_ADMIN. Another option would be
to load the XDP program using libbpf.