old-main can quit before informing main
When during a reload, old-main has finished all it's tasks (killing the old server processes, informing xfrd that it's going to exit etc.), it will shutdown (cmd == NSD_QUIT), but before it shuts down, it informs the reload process that it is done by sending it NSD_RELOAD over the reload_listerener.
However, on some platforms, the reload_listener pipe may be closed (by old-main's) shutdown, before the reload process can read from it.
This was nog logged as an error before (NSD 4.8) and neither should it now.